+++ src/pdf-page-labels.c
@@ -44,7 +44,7 @@ int pdfout_page_labels_set (fz_context *ctx, pdf_document *doc,
       pdf_obj *dict_obj, *string_obj;
       pdfout_page_labels_style_t style;
       char *text, *prefix;
-      size_t text_len;
+      int text_len;
       pdfout_page_labels_mapping_t *mapping;
 
       mapping = pdfout_page_labels_get_mapping (labels, i);
@@ -539,7 +539,7 @@ page_labels_get (pdfout_page_labels_t *labels, fz_context *ctx,
 	      int pdf_buf_len = pdf_to_str_len (ctx, object);
 	      if (pdf_buf_len)
 		{
-		  size_t utf8_buf_len;
+		  int utf8_buf_len;
 		  mapping.prefix = pdfout_pdf_to_utf8 (ctx, pdf_buf,
 						       pdf_buf_len,
 						       &utf8_buf_len);
